Change logweasel to a function that does not track state
Because LW uses a singleton to track initialization state (the
application name), any later initializations of LW can overwrite the
original initialization values. Now that we are distributing components
with LW in them, we are guaranteed to have multiple initializations. We
already have 3 inits that overwrite themselves in KIUI – the app itself,
ER, and the Graph API helper. So as the app runs, the supposedly
universal app name is changed three times.

Instead of enabling spooky action at a distance, we’ll make log weasel a
simple function with no initialization. If people want to standardize
the name they’re passing in they can either pull it from a single source
or make a wrapping function that contains a constant. This is a little
more work but allows us to more easily trace xhr requests back to what
called them.